# Update resvg to 0.34 for improved performance