--- a/vl-convert-rs/Cargo.toml
+++ b/vl-convert-rs/Cargo.toml
@@ -21,11 +21,11 @@ tokio = {version= "1.21", features=["rt-multi-thread"]}
 reqwest = {version="=0.11.12", default-features=false, features=["rustls-tls", "blocking"]}
 rustls = "=0.20.6"
 lazy_static = "1.4.0"
-usvg = "0.32.0"
-resvg = "0.32.0"
-tiny-skia = "0.9.0"
+usvg = "0.33.0"
+resvg = "0.33.0"
+tiny-skia = "0.9.1"
 png = "=0.17.6"
-fontdb = {version="0.13.0", features=["fontconfig"]}
+fontdb = {version= "0.14.1", features=["fontconfig"]}
 tempfile = "3.3.0"
 log = "0.4.17"
 env_logger = "0.10.0"