--- a/vl-convert-rs/Cargo.toml
+++ b/vl-convert-rs/Cargo.toml
@@ -34,6 +34,7 @@ http = "0.2.8"
 [dev-dependencies]
 tokio = {version="1.21", features=["macros", "rt", "rt-multi-thread"]}
 rstest = "0.17.0"
+dssim = "3.2.4"
 
 [build-dependencies]
 serde_json = "1.0.85"