… Cargo.lock
… thirdparty_rust.yaml
… vl-convert-python/thirdparty_rust.yaml
… vl-convert-rs/tests/vl-specs/expected/v5_8/bar_chart_trellis_compact-excel.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/bar_chart_trellis_compact.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/circle_binned-dark.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/circle_binned.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/font_with_quotes.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/line_with_log_scale-fivethirtyeight.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/line_with_log_scale.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/maptile_background.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/remote_images.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/stacked_bar_h-vox.png
… vl-convert-rs/tests/vl-specs/expected/v5_8/stacked_bar_h.png
… vl-convert-rs/thirdparty_rust.yaml
… vl-convert/thirdparty_rust.yaml